Reorganize classes between 'reading' and 'writing' sub-packages
This update moves classes related to 'Result' to the 'reading' sub-package, and those related to 'ManipulationQuery' are shifted to 'writing' for better context separation. It also enhances code readability through added comments, renaming of classes and other significant refactorings.
